many changes to the process of obtaining prices
note: Need to rewrite the entire process to simplify it and have less "scraping".
Also, need to simplify an order in which we determine what type of prices we want to store, e.g. NM only, what happens if there's only SP in stock, what happens if theres no buylist whatsoever?